+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * CHANGED: Renamed PublicSuffixService::RuleList to PublicSuffixService::List.
 
+* CHANGED: Renamed PublicSuffixService::List#list to PublicSuffixService::List#rules.
+
 
 ## Release 0.9.0